requirements.txt
RPi.GPIO
mqtt_client
